Create marsyasSegmenter.cpp
The marsystem component "SelfSimilarityMatrix" can have no more than one child, to overpass the mandatory Sourcefile and Gain in this example they are inclosed in different series. In order to calculate the SelfSimilarityMatrix of a single feacture vector we add the spectrum to a Series and this series to the SelfSimilarityMatrix MarSystem.